+++ lib/macos_window_toolkit.dart
@@ -1,6 +1,7 @@
 library;
 
 export 'src/macos_window_toolkit.dart';
+export 'src/models/macos_screen_info.dart';
 export 'src/models/macos_version_info.dart';
 export 'src/models/macos_window_info.dart';
 export 'src/models/platform_error_code.dart';